UI Test Cleanup: Extract match_ref_pats tests
